% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{autocor}
\alias{autocor}
\title{Calculate the circular autocorrelation in a time-series}
\usage{
autocor(ts)
}
\arguments{
\item{ts}{A time-series (a numeric vector)}
}
\value{
A scalar numerical value circular autocorrelation
}
\description{
Calculate the circular autocorrelation in a time-series
}
\details{
Note that all values are included. Non-finite values may cause odd answers
}

