% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePower.r
\name{calculatePower}
\alias{calculatePower}
\title{Estimate the power of the random thinning test based on the properties a given time-series}
\usage{
calculatePower(ts, p, nMC = 1000, nThin = 1000, its = NULL,
  fr = c(0.001, 0.025, 0.01, 0.025, 0.1), pThreshold = 0.05,
  useNApattern = FALSE)
}
\arguments{
\item{ts}{A time-series (a numeric vector)}

\item{p}{Length of the period to test for (a positive integer)}

\item{nMC}{Number of samples to estimate the power (a positive integer)}

\item{nThin}{Number of resamples to use in the random thinning test (a positive integer)}

\item{its}{Indices of which bin the values of x correspond to (a numeric vector of integers between 1 and p). This is optional, and taken to be cyclical by default.}

\item{fr}{Which fractions to consider discarding (a numeric vector, with values between 0 and 1, although it is best if they are kept between between 0.001 and 0.2)}

\item{pThreshold}{P-value threshold to use when calculating the power (a positive scalar between 0 and 1, although it is recommended to be around 0.05)}

\item{useNApattern}{Logical value - if \code{TRUE}, the synthetic time-series inherit the same missingness pattern as in the original time-series}
}
\value{
A numeric vector, with one entry per value of \code{fr}, giving the power of the test for this particular value
}
\description{
Estimate the power of the random thinning test based on the properties a given time-series
}
\examples{
## Generate a random time-series:
##  - length = 1000
##  - signal period = 8
##  - signal to noise ratio = 0.2
set.seed(42)

x <- rep(rnorm(8)*0.2,length.out = 1000) + rnorm(1000)
calculatePower(ts = x, p = 8, nMC = 200, nThin = 200,
                           fr = c(0.01,0.025,0.1))
}

