% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTSstatistics.r
\name{calculateTSstatistics}
\alias{calculateTSstatistics}
\title{Calculate various statistics for the time-series provided as an argument}
\usage{
calculateTSstatistics(ts, its = NULL, p)
}
\arguments{
\item{ts}{A time-series (a numeric vector)}

\item{its}{Indices of which bin the values of x correspond to (a numeric vector of integers between 1 and p). This is optional, and taken to be cyclical by default.}

\item{p}{Length of the period to test for (a positive integer)}
}
\value{
A list with the following components:
 - \code{signalNoiseRatio}: The signal to noise ratio estimated for the time-series
 - \code{noiseAC}: The autocorrelation of the noise time-series
 - \code{signalAC}: The (circular) autocorrelation of the signal of length \code{p}
 - \code{nSignals}: The number of length \code{p} periods in the time-series
}
\description{
Calculate various statistics for the time-series provided as an argument
}
\examples{
## Generate a random time-series:
##  - length = 1000
##  - signal period = 8
##  - signal to noise ratio = 0.2
set.seed(42)

x <- rep(rnorm(8)*0.2,length.out = 1000) + rnorm(1000)
calculateTSstatistics(ts = x, p = 8)
}

