% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateRandomSignal}
\alias{generateRandomSignal}
\title{Generate a random signal with given circular autocorrelation}
\usage{
generateRandomSignal(m, sac, maxcounter, margin)
}
\arguments{
\item{m}{Length of the signal (a positive integer)}

\item{sac}{Strength of the circular autocorrelation (a positive double, should be between 0 and 1)}

\item{maxcounter}{Maximum number of tries in the \code{while} loop (a positive integer)}

\item{margin}{The margin of error between the desired and obtained autocorrelation (a positive double)}
}
\value{
A numerical vector
}
\description{
Generate a random signal with given circular autocorrelation
}

