% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rangeTestFinitesOnly}
\alias{rangeTestFinitesOnly}
\title{Range test - optimized for finite values only}
\usage{
rangeTestFinitesOnly(ts, its, p)
}
\arguments{
\item{ts}{A time-series (a numeric vector)}

\item{its}{Indices of which bin the values of x correspond to (a numeric vector of integers between 1 and p). This is optional, and taken to be cyclical by default.}

\item{p}{Length of the period to test for (a positive integer)}
}
\value{
A scalar numerical value giving the maximum difference (maximum - minumum) of the means of the \code{p} bins
}
\description{
Range test - optimized for finite values only
}

