% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rangeTestGeneral}
\alias{rangeTestGeneral}
\title{Range test - general version}
\usage{
rangeTestGeneral(ts, its, p)
}
\arguments{
\item{ts}{A time-series (a numeric vector)}

\item{its}{Indices of which bin the values of x correspond to (a numeric vector of integers between 1 and p). This is optional, and taken to be cyclical by default.}

\item{p}{Length of the period to test for (a positive integer)}
}
\value{
A list with three components:
  - \code{range} = The maximum difference (maximum - minumum) of the means of the \code{p} bins
  - \code{counts} = a numerical array giving number of values in each of the \code{p} bins
  - \code{means} = a numerical array giving the mean value in each of the \code{p} bins
}
\description{
Range test - general version
}

